refactor(npu): isolate DSV4 Async CAM forward
